Added a new parameter to ComServer object constructor
The main problem we currently have to properly use node-dcom with
newer versions of Windows is related to a header problem. We use
an older DCOM version number which is causing 'Unknown Error' when
we try to connect. This should solve the problem for most systems
with apropriate DCOM configuration. Now when you create a ComServer
you should also provide an object containing the major and minor DCOM
version number to be used. The default value uses will always be 5.4
but for most Windows systems the current version is 5.7 .